removed code for FirstScheduler and refactored SecondScheduler to BackTrackingScheduler with interface Scheduler.
