+++ src/main/java/com/example/demo/services/TimeTableService.java
@@ -8,7 +8,7 @@ import com.example.demo.models.enums.Semester;
 import com.example.demo.models.enums.Status;
 import com.example.demo.repositories.TimeTableRepository;
 import com.example.demo.scheduling.Scheduler;
-import com.example.demo.scheduling.SecondScheduler;
+import com.example.demo.scheduling.BacktrackingScheduler;
 import jakarta.persistence.EntityNotFoundException;
 import jakarta.transaction.Transactional;
 import org.slf4j.Logger;
@@ -38,7 +38,7 @@ public class TimeTableService {
     private static final Logger log = LoggerFactory.getLogger(TimeTableService.class);
 
     public TimeTableService(TimeTableRepository timeTableRepository, RoomTableService roomTableService,
-                            CourseSessionService courseSessionService, DTOConverter dtoConverter, SecondScheduler scheduler,
+                            CourseSessionService courseSessionService, DTOConverter dtoConverter, BacktrackingScheduler scheduler,
                             GlobalTableChangeService globalTableChangeService) {
         this.timeTableRepository = timeTableRepository;
         this.roomTableService = roomTableService;
@@ -221,12 +221,12 @@ public class TimeTableService {
             return false;
         }
     }
-    /* TODO: fix
+
     @PreAuthorize("hasAnyAuthority('ADMIN', 'USER')")
     public List<CourseSession> checkCollisions(TimeTable timeTable){
         return scheduler.collisionCheck(timeTable);
     }
-    */
+
     @PreAuthorize("hasAnyAuthority('ADMIN', 'USER')")
     public TimeTable unassignAllCourseSessions(TimeTable timeTable){
         courseSessionService.unassignCourseSessions(timeTable.getAssignedCourseSessions());
